refactor(components.expiration): move to webUniverseComponents
