--- client/app/components/expiration/service-expiration-date.component.css
@@ -1,11 +1,0 @@
-.expiration .expired {
-    color: #b94a48 !important;
-}
-
-.expiration .expired7 {
-    color: #f60 !important;
-}
-
-.expiration .expired15 {
-    color: #c0ac00 !important;
-}
