--- client/app/components/expiration/service-expiration-date.component.js
@@ -1,10 +1,0 @@
-angular.module('directives').component('serviceExpirationDate', {
-  templateUrl: 'components/expiration/service-expiration-date.component.html',
-  bindings: {
-    serviceInfos: '<',
-    hideRenewAction: '<',
-    serviceType: '@',
-    serviceName: '<',
-  },
-  controller: 'ServiceExpirationDateComponentCtrl',
-});
