+++ client/app/core/translations/Messages_fr_FR.xml
@@ -110,7 +110,6 @@
    <translation id="wizard_close" qtlid="4670">Fermer</translation>
    <translation id="wizard_order" qtlid="43558">Commander</translation>
    <translation id="wizard_pay" qtlid="130183">Régler</translation>
-   <translation id="wizard_renew" qtlid="118962">Renouveler</translation>
    <translation id="wizard_next" qtlid="17743">Suivant</translation>
    <translation id="wizard_previous" qtlid="17431">Précédent</translation>
    <translation id="wizard_export" qtlid="69881">Exporter</translation>
@@ -562,11 +561,6 @@
    <translation id="core_change_owner_error" qtlid="270735">Impossible d’obtenir les informations necessaires au changement de propriétaire</translation>
    <translation id="contacts_management" qtlid="32206">Contacts</translation>
    <translation id="more_tabs" qtlid="270748">Autres onglets</translation>
-   <translation id="expired_date" qtlid="382076">Expiré le <strong>{0}</strong></translation>
-   <translation id="expiration_date" qtlid="440274">À renouveler manuellement le <strong>{0}</strong></translation>
-   <translation id="resiliation_date" qtlid="382089">Résiliation prévue le <strong>{0}</strong></translation>
-   <translation id="resiliation_cancel_label" qtlid="285350">Annuler la résiliation</translation>
-   <translation id="renew_date" qtlid="440287">Renouvellement automatique prévu en <strong>{0}</strong></translation>
    <translation id="core_email_voila_title" qtlid="271751">Adresse @voila.fr</translation>
    <translation id="core_email_voila_alert" qtlid="271764">Du fait de la fermeture annoncée du service Voila Mail, toutes les adresses @voila.fr seront désactivées à compter du 12 janvier 2016. Les possesseurs d'adresses en @voila.fr devront donc se tourner vers un nouveau webmail.</translation>
    <translation id="core_email_voila_alert2" qtlid="271777">Pour continuer à recevoir nos informations par e-mail, nous vous conseillons de renseigner dès à présent une nouvelle adresse de contact dans la section Mon compte.</translation>
