+++ client/app/css/source.less
@@ -59,7 +59,6 @@
 @import (less) "@ovh-ux/ovh-utils-angular/bin/template/wizard/wizardStep/wizardStep.css";
 @import (less) "font-awesome/css/font-awesome.min.css";
 @import (less) "animate.css/animate.min.css";
-@import (less) "../components/expiration/service-expiration-date.component.css";
 @import (less) "@bower_components/angular-xeditable/dist/css/xeditable.min.css";
 @import (less) "../components/fileEditor/file-editor.css";
 @import (less) "ovh-angular-sso-auth-modal-plugin/dist/ovh-angular-sso-auth-modal-plugin.min.css";
