+++ client/app/hosting/hosting.html
@@ -50,12 +50,12 @@
                               data-ng-message="maxlength"></span>
                     </div>
                 </div>
-                <service-expiration-date
+                <wuc-service-expiration-date
                         data-ng-if="hosting.serviceName && hosting.serviceInfos"
                         service-name="hosting.serviceName"
                         service-type="HOSTING_WEB"
                         service-infos="hosting.serviceInfos">
-                </service-expiration-date>
+                </wuc-service-expiration-date>
             </form>
         </div>
         <div class="col-sm-3 d-flex">
