+++ client/app/private-database/private-database.html
@@ -59,11 +59,11 @@
                 </div>
 
                 <div class="mt-4" data-ng-if="database.serviceName && database.serviceInfos">
-                    <service-expiration-date
+                    <wuc-service-expiration-date
                         service-name="database.serviceName"
                         service-type="HOSTING_PRIVATE_DATABASE"
                         service-infos="database.serviceInfos">
-                    </service-expiration-date>
+                    </wuc-service-expiration-date>
                 </div>
             </div>
