+++ client/app/webUniverseComponents/expiration/service-expiration-date.component.js
@@ -1,0 +1,12 @@
+import template from './service-expiration-date.component.html';
+
+export default {
+  template,
+  bindings: {
+    serviceInfos: '<',
+    hideRenewAction: '<',
+    serviceType: '@',
+    serviceName: '<',
+  },
+  controller: 'WucServiceExpirationDateComponentCtrl',
+};
