+++ client/app/webUniverseComponents/expiration/service-expiration-date.component.less
@@ -1,0 +1,13 @@
+wuc-service-expiration-date { /* stylelint-disable-line */
+  .expiration .expired {
+    color: #b94a48 !important;
+  }
+
+  .expiration .expired7 {
+    color: #f60 !important;
+  }
+
+  .expiration .expired15 {
+    color: #c0ac00 !important;
+  }
+}
