+++ client/app/webUniverseComponents/expiration/translations/Messages_fr_FR.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<translations>
+   <translation id="expiration_expired_date" qtlid="382076">Expiré le <strong>{0}</strong></translation>
+   <translation id="expiration_renew_date" qtlid="440287">Renouvellement automatique prévu en <strong>{0}</strong></translation>
+   <translation id="expiration_resiliation_cancel_label" qtlid="285350">Annuler la résiliation</translation>
+   <translation id="expiration_wizard_renew" qtlid="118962">Renouveler</translation>
+   <translation id="expiration_resiliation_date" qtlid="382089">Résiliation prévue le <strong>{0}</strong></translation>
+   <translation id="expiration_expiration_date" qtlid="440274">À renouveler manuellement le <strong>{0}</strong></translation>
+   <translation id="expiration_resiliation_cancel_label" qtlid="285350">Annuler la résiliation</translation>
+</translations>
