+++ client/app/webUniverseComponents/index.js
@@ -3,6 +3,7 @@ import angular from 'angular';
 import wucConverter from './converter';
 import wucCron from './cron';
 import wucCronValidator from './cron-validator';
+import wucExpiration from './expiration';
 import wucFileChange from './fileChange';
 import wucProgressBarElementCounter from './progressBarElementCounter';
 
@@ -13,6 +14,7 @@ angular
     wucConverter,
     wucCron,
     wucCronValidator,
+    wucExpiration,
     wucFileChange,
     wucProgressBarElementCounter,
   ]);
